Realizacion de request y response en la carpeta commons

+++ GESTION DE HOTELES/msv-reservas/src/main/java/com/hotel/reservas/models/Reserva.java
@@ -46,8 +46,8 @@ public class Reserva {
 
 	@NotNull(message = "Su total prro, vuelva pronto")
     @Column(name = "total", precision = 10, scale = 2)
-    private BigDecimal total;
-
+    private Double total;
+ 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private EstadoReserva estado = EstadoReserva.CONFIRMADA;
